+++ partnerProject/file.h
@@ -5,7 +5,7 @@
 void fileFunction() {
 	std::cout << "This is the file printing a message!\n";
 }
-//Adds to an inputted float
-float addToNum(float num) {
-	return num + 3.0f;
+//Adds to an inputted int
+int addToNum(int num) {
+	return num + 3;
 }
